Fix second round starting from player two bug
Inside of handleTurn (previously placeMark) playing a turn
comes before handling a win. And because playing a turn changed
the current player at the end, declaring the winner from inside
handleWin was impossible. It always declared the oppisite
player. By declaring a new const player for each click on cell,
I was able to save the current player and declare him from
the handleWin method, by using player as an argument.